requirements.txt
asgiref==3.6.0
CacheControl==0.12.11
cachetools==5.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
colorama==0.4.6
cryptography==39.0.0
Deprecated==1.2.14
Django==4.1.5
django-minio-storage==0.5.7
djangorestframework==3.14.0
et_xmlfile==2.0.0
firebase-admin==6.0.1
fpdf==1.7.2
gcloud==0.18.3
google-api-core==2.11.0
google-api-python-client==2.73.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.9.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.51.1
gunicorn==21.2.0
httplib2==0.21.0
idna==3.4
jwcrypto==1.5.1
minio==7.1.17
msgpack==1.0.4
numpy==2.2.0
oauth2client==4.1.3
openpyxl==3.1.5
packaging==23.2
panda==0.3.1
pandas==2.2.3
Pillow==10.1.0
proto-plus==1.22.2
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.16.0
pycryptodomex==3.20.0
PyJWT==2.6.0
pyparsing==3.0.9
pypdf==3.16.4
Pyrebase4==4.7.1
python-dateutil==2.9.0.post0
python-jwt==4.1.0
pytz==2022.7.1
qrcode==7.3.1
reportlab==4.0.6
requests==2.28.2
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
wrapt==1.16.0
